marketplace_bot (copy)/requirements.txt
aiogram==3.4.1
openpyxl
python-dotenv
splite3
